@@ -1,4 +1,4 @@
-class CreateTags < ActiveRecord::Migration
+class CreateTags < ActiveRecord::Migration[6.0]
   def change
     create_table :tags do |t|
       t.string :name
