@@ -15,7 +15,7 @@ Rails.application.configure do
   # Configure public file server for tests with Cache-Control for performance.
   config.public_file_server.enabled = true
   config.public_file_server.headers = {
-    'Cache-Control' => 'public, max-age=3600'
+    'Cache-Control' => "public, max-age=#{1.hour.to_i}"
   }
 
   # Show full error reports and disable caching.
@@ -26,8 +26,7 @@ Rails.application.configure do
   config.action_dispatch.show_exceptions = false
 
   # Disable request forgery protection in test environment.
-  # config.action_controller.allow_forgery_protection = false
-  # config.action_mailer.perform_caching = false
+  config.action_controller.allow_forgery_protection = false
 
   # Tell Action Mailer not to deliver emails to the real world.
   # The :test delivery method accumulates sent emails in the
@@ -35,18 +34,8 @@ Rails.application.configure do
   config.action_mailer.delivery_method = :test
 
   # Print deprecation notices to the stderr.
-  config.active_support.deprecation = :raise # :stderr #:log
-
-  # Raise an exception on deprecations for aggressive deprecation cleanups
-  ActiveSupport::Deprecation.behavior = :raise
-  # ActiveSupport::Deprecation.silenced = true
+  config.active_support.deprecation = :stderr
 
   # Raises error for missing translations
-  config.action_view.raise_on_missing_translations = true
-
-  # Randomize the order test cases are executed.
-  config.active_support.test_order = :random # :sorted
-
-  # Set to :debug to see everything in the log.
-  config.log_level = :fatal # :debug # :info
+  # config.action_view.raise_on_missing_translations = true
 end
